mononoke_api: add no-op file changes check to create_changeset/create_changeset_stack
Summary: Add another check in `create_changeset`/`create_changeset_stack` `mononoke_api` method to prevent the creation of no-op files changes i.e. file changes that don't change the content of the file. The check is gated behind a parameter that users can supply, this is because some users might not be able to handle the extra latency resulting from depending on `Fsnodes`. Might make sense to gate skipping the check behind an ACL later on.

Reviewed By: RajivTS

Differential Revision: D84351712

fbshipit-source-id: 7bead9a53d765a8f254008bad9a20bbbc67afbfe