TreeInode: add option to skip writing directories to overlay on read
Summary:
When the new experimental.lazy-inode-persistence config is set to true, eden will no longer eagerly write directories to the overlay on read. Instead, non-materialized directories are written to the overlay when unloaded, except during checkout. Note that checkout only unloads inodes with fs-ref-count==0, so open or cached inode numbers will still be remembered in memory.

There are two main reason to remember inode numbers:
- For files still referenced by the fs, we obviously need to remember inode numbers since FUSE requests will reference those inode numbers.
- For files not referenced by the fs, some programs remember inode numbers, and then use them later to check if a file is identical. The one example we know is clang when evaluating `#pragma once` directives. In general it seems a reasonable expectation that a filesystem's inode numbers won't "randomly" change.

With lazy-inode-persistence=true, we don't persist inodes on checkout because we don't want the user to wait while we write thousands of directories to the overlay. In general, "checkout" seems like the perfect time to "break" certain rules to optimize performance. I don't think anyone expects a clang build to work across a "checkout".

Of course, now if eden crashes we will switch inode numbers on restart. The clang build isn't going to work anyway if eden crashes, so I think this is okay.

The always-write-directories-to-overlay behavior was added in D7598094 which references race conditions related to inode metadata being keyed by inode number. I'm not sure this is an issue, though, since changing inode metadata should materialize everything (right?), so we don't need to pre-materialize directories.

I enabled the config flag by default for integration tests. There are some good tests for inode number stickiness, such as persistence_test.py.

Reviewed By: genevievehelsel

Differential Revision: D84164596

fbshipit-source-id: 078637ce0bc13718b972daf0c4b761ffc0b14d42