add stopAllGarbageCollections to the eden shutdown process
Summary:
## This diff:
Add `stopAllGarbageCollections` to the eden shutdown process. In most cases (more than 99% based on the crash rate in scuba) there is no running GC during eden shutdown and this function returns true immediately. However, if a GC is running, it sends the cancellation token and then check if it exited in a loop for three times.

## Context:
It is very rare that GC and eden shutdown happen at the same time. However, if these two happen at the same time the shutting down eden will crash instead of shutdown. The GC is running inside the ThreadPool and there is no direct way to force-stop a ThreadPool task. We should send the cancellation token to the task, and request cancellation from the token.

Differential Revision: D84110705

fbshipit-source-id: e0db438151b60d9103a3711fc426b8bd3b7f919e