move stopAllGarbageCollections logic into one function
Summary:
## This diff:
No logic change. I just want to move the GC stopping logic into a separated function then I can reuse it in other places. I decrease the `retryInterval` from 20 seconds per iteration to 1 second per iteration. In this case the function can block the eden shutdown for a maximum of 3 seconds and then crash if the GC doesn't stop. Running GC and eden shutdown at the same time is very rare situation and if it happens in most cases GC will stop in the first try, then we don't need to wait for 20 seconds here.

## Context:
It is very rare that GC and eden shutdown happen at the same time. However, if these two happen at the same time the shutting down eden will crash instead of shutdown. The GC is running inside the ThreadPool and there is no direct way to force-stop a ThreadPool task. We should send the cancellation token to the task, and request cancellation from the token.

Differential Revision: D84107746

fbshipit-source-id: 18ad4e7e7da384b5c827bf0c920b2b3ecf9b34e0