add more cancellation token checks to GC
Summary:
## This diff:
Add more cancellation token checks to GC. Then we can confirm that GC stop sooner when we request a cancellation through token.

## Context:
It is very rare that GC and eden shutdown happen at the same time. However, if these two happen at the same time the shutting down eden will crash instead of shutdown. The GC is running inside the ThreadPool and there is no direct way to force-stop a ThreadPool task. We should send the cancellation token to the task, and request cancellation from the token.

Differential Revision: D84118938

fbshipit-source-id: 56e6acf653eb3a0c5e4a31484cc44ccfa5e20874