create_repos: Make it retriable on failure to land the mutation
Summary:
While there are situations where manual recovery will be necessary (for instance, if the tw tasks serving a `create_repos` request restarts while processing the request), the most common issue likely to fail a repo creation would be a failure to land the mutation.

This can happen for at least a couple of reasons:
* Canary cancelled by user
* Conflict with another configerator change

In these situations, without this change, trying to create the same repository again would fail as the db would contain an entry such as:
```
+-------+---------+---------------+-----------------+
| id    | repo_id | repo_name     | source_of_truth |
+-------+---------+---------------+-----------------+
| 17227 |   17736 | pierrec-test2 | reserved        |
+-------+---------+---------------+-----------------+
```
and the next invocation to `create_repos` would try to insert a new row conflicting with this one (the `repo_name` is unique in that table).

Since the mutation failed to land, it is fail to "free" the reserved repo and make the command retriable.

Do it.

Reviewed By: clara-9

Differential Revision: D84360325

fbshipit-source-id: f2b45d7ed7296df343bd6ea79a324ec7350ea96c