Share config handle among all repos
Summary:
Previously we were loading the configuration handle for every single repo we loaded. This change will load it only once and share it among all other repos.

In the diff_service, which needs to load all repos (around 2k per task, with the current deployment), we are having memory problem, with several tasks OOMing during loading, and after profiling the memory usage, we could see that around 80% of memory usage was due to this.

Reviewed By: clara-9

Differential Revision: D84266221

fbshipit-source-id: c138ab5f7ec255f6503b21fb08eff7733980dd26