Diff Router: JK selection between local and remote diffs
Summary:
Add a new JK to control which version to use for diffs, the current version using XDiff directly, or the remote diff server.

A coming diff adds a new config parameter to select between one option or the other, and another one to use the new diff feature crate (used in the diff_service) to replace the direct use of xdiff and avoid code duplication.

Reviewed By: markbt

Differential Revision: D83972801

fbshipit-source-id: 775b17c0c0130219603eb30bb68b8973468c13cc