remove extra heartbeat files on startup or clean exit
Summary: When I was playing with silent exit, I found that there could be some rare edge cases that we have multiple heartbeat files in the eden directory. We should confirm that we only get log for the first heartbeat file and the rest are removed without getting log. Otherwise we will have multiple logs per one crash.

Reviewed By: muirdm

Differential Revision: D84398022

fbshipit-source-id: 0e1806e5f9952f5c9bb150ed54e7c87d4e130aff