fix async silent daemon exit logger
Summary:
Inside the `HeartbeatManager` class, we are making a shared pointer from this class to pass to a lambda function.
https://www.internalfb.com/code/fbsource/[f3b95e0e636a]/fbcode/eden/fs/service/HeartbeatManager.cpp?lines=152
But the `HeartbeatManager` object was defined as a unique pointer in the `EdenServer`. This discrepancy resulted in a failure in this threadpool task:
https://www.internalfb.com/code/fbsource/[f3b95e0e636a]/fbcode/eden/fs/service/HeartbeatManager.cpp?lines=150-156
This failure didn't break anything but the silent exit logger which was inside this Threadpool task didn't work.

Fix the object definition in `EdenServer` to fix this issue and get the silent exit logger data back

Reviewed By: muirdm

Differential Revision: D84395820

fbshipit-source-id: 92a4a42e80201a9f7e4573fa595fbd70d4ab23f2