purge: explictly warn on unsupported submodules
Summary:
Give explicit warning that submodules are not supported for purge, so users can remove on their own.

Currently submodules to be purged/cleaned are attempted with unlink() just like files, and got OSErrors (EISDIR) eventaully since they're directories.

The usecase should be rare for submodule users. Even if an untracked submodule can get removed correctly,  the new entry in .gitmodules still needs to be manually cleaned up. The unhandled side effects can be surprising to users without knowledge of submodules who simply wants to undo everything. As a result, the actual support will need extra complexity to be introduced into purge/clean implementation, which is not ideal for now.

Reviewed By: muirdm

Differential Revision: D84274551

fbshipit-source-id: 2da6e2ea3fa54701e9c31498bb19942ba8e547b7