add a EphemeralFsInodeCatalog
Summary:
This diff add a new option to run a "Legacy" overlay in a oneshot mode. In this hybrid mode, we run with a in-memory inode catalog (directories) and on-disk file content storage (files).

The code for `EphemeralFsInodeCatalog` is pretty much the same as `MemInodeCatalog` except it manages a `FSFileContentStore*` for open/close and it doesn't implement some of the "optional" FSCK methods that we need for `MemInodeCatalog` on Windows. If this proves successful, we can combine `MemInodeCatalog` and `EphemeralFsInodeCatalog`, I wanted to avoid any behavior changes on Windows for this experiment so I semi-forked the code.

This is **NOT** safe to run in environments where users expect work to be saved. If EdenFS crashes, we cannot easily recover data.

I have yet to test this with graceful restart. I wouldn't expect it to work at all. I'd expect what would happen is that the restart would work but no mounts would be mounted.

Reviewed By: muirdm

Differential Revision: D84524846

fbshipit-source-id: 52713c6a3e5619d7a313342ae5deb25a3bce30d9