admin/derivation_queue: display all needed items not just ready items
Summary: Currently when using the `monad derivation-queue summary` command we display the total number of needed items, but then all display information for the items which are ready for derivation in the table. This diff modifies the logic to display all items, and adds an `is_ready` column to be able to till apart items which don't have any outstanding dependencies and are ready to be picked up by a derivation worker.

Reviewed By: mzr

Differential Revision: D84604986

fbshipit-source-id: 2044427e733065576ae6e4127d8f069e8ab759d8