Add cancellation source map
Summary:
In this diff, add an in memory map relating a Thrift RequestID to a cancellation source.

This will be the source of truth for triggering cancellations.

`INSTRUMENT_THRIFT_CALL` will be responsible for adding its own request ID to the map during construction, and removing its own request id during destruction

Reviewed By: kavehahmadi60

Differential Revision: D83778481

fbshipit-source-id: 13cba0590cc6f6327ef34c7e00d5b91de15485e2