edensparse: teach cat to respect edensparse filters
Summary:
# This diff

Wraps the cat function to ensure the current active edensparse filter is respected when returning output. We achieve this by using an intersectmatcher when edensparse is in use.

# Context

Some commands (such as 'sl cat' and 'sl grep') do not currently respect filters. We'd like to teach them how to respect filters so that users aren't confused by the results.

This stack adds some tests and implements filtering for 'sl cat' and 'sl grep'

Reviewed By: muirdm

Differential Revision: D84401240

fbshipit-source-id: b8859c1292a45a6e76054cfc46e4b324efe6d8a7