integration: add 'hg grep' tests
Summary:
# This diff

Adds tests that show the current behavior of `sl grep` when filters are enabled.

NOTE: you'll see that it appears that 'sl grep' respects filters. While this is partially true, it doesn't respect filters when biggrep is used. Since our integration tests don't exercise that codepath, it's not possible to show the bad behavior in this diff.

# Context

Some commands (such as 'sl cat' and 'sl grep') do not currently respect filters. We'd like to teach them how to respect filters so that users aren't confused by the results.

This stack adds some tests and implements filtering for 'sl cat' and 'sl grep'

Reviewed By: muirdm

Differential Revision: D84401239

fbshipit-source-id: 72d7010835b9072b9ebe13cc18df23744be54430