edensparse: teach grep to respect edensparse filters
Summary:
# This diff

Teaches grep to respect filters by wrapping the grep command and passing in an intersectmatcher that requires both the sparse filter and the passed pats option to be satisfied

# Context

Some commands (such as 'sl cat' and 'sl grep') do not currently respect filters. We'd like to teach them how to respect filters so that users aren't confused by the results.

This stack adds some tests and implements filtering for 'sl cat' and 'sl grep'

Reviewed By: muirdm

Differential Revision: D84401238

fbshipit-source-id: 8a0f87c7085807d884e1aa7019b2647b8e0f9846