Fix inferred_copy_from taking path prefix
Summary:
I'm getting this error when backfilling inferred_copy_from for fbsource.

We have a slightly different derived data config for this type for fbsource -- when we do basename matching, we search the top 2 level directory instead of 1. The path prefix extraction is failing for commits that touched the top level files.

Current failure:
```
[lyang@devvm32214.prn0 ~/configerator (d3bcc33fd3)]$ monad --prod derived-data -R fbsource derive --derived-data-types inferred_copy_from -i 739567f9ec539ea56cea86885b3f5441fea9a2f58c099b71ae55dfcbafd3813e --from-predecessor
Error: taking 2 components but path only has 1
```

Reviewed By: YousefSalama

Differential Revision: D84809418

fbshipit-source-id: a4ab5730100630d896a7965976bdc544eaab9374