Ensure that error to upload LFS files before git push results in a user error
Summary:
In the current implementation, Mononoke Git server ends up reporting a system failure when the user tries to push content to a Git repo that contains LFS pointers that are not uploaded to the Git LFS store.

This has two unwanted consequences:
- Our push SLI gets affected without reason because the underlying problem in this case is a user / human error that needs to be rectified
- Users have no idea of what they did wrong because the error simply shows up as HTTP 500 status code on the Git client which leads to confusion and support posts in the Source Control group

Ideally, we want to avoid both. Ensure that such errors are not counted as reliability dips and also provide enough context to the user so they know what they did wrong.

NOTE: There was a flaw in the error message reporting logic which I had to fix and as a consequence two more integration tests had to be updated.

Reviewed By: YousefSalama

Differential Revision: D84702642

fbshipit-source-id: ac01e7cd21181245b402bda836d9541330e0f631