Pass review scope to Devmate and update banner text
Summary:
Right now, if you have a stack of local commits and some uncommitted changes, it's a bit confusing as to what Devmate will review when you click the ISL upsell. To help with this, we can do 2 things:

1. Adjust the content of the banner depending on the context (whether we're in commit/amend mode, whether there are uncommitted changes)
2. Pass the review "scope" as context to the model prompt to (hopefully) influence to use the correct local changes. I thought this could be an easy way to pass the required context.

I also just disable the button now when we're in commit mode and there are no uncommitted changes (since there's nothing to review), and I added a tooltip to the button to explain why it's disabled.

This did uncover that we weren't even showing the upsell in commit mode, so we should get a discoverability improvement as well.

I also make similar changes for (2) in the smart actions (we're already changing the text)

Differential Revision: D84374807

fbshipit-source-id: c792f72bb940dd991f2bf7d294b84eb94dba782c