edenfs-client: stop populating deprecated filterId field
Summary:
# This diff

Makes the Sapling client stop passing the filterId field to EdenFS (since the daemon will ignore it)

# Context

It's not valid to populate Thrift response fields of type "String" with invalid UTF-8. This causes an error in the Thrift layer when the result is passed to Rust/Python.

Therefore, we should stop populating the deprecated field with String type and only pass the field with "bytes" type.

No downstream clients rely solely on the deprecated String field, so it's okay to remove handling of it entirely from the daemon. At the same time, we can make all clients stop populating/reading this field as well.

Note: In a couple weeks, we can remove this field from the thrift spec entirely.

Reviewed By: muirdm

Differential Revision: D84760898

fbshipit-source-id: ea7e21fd1be6f70f334db2dc247e13129d11073a