cli: kill Python version of du
Summary:
# This diff

Completely kills the Python version of `eden du`, but leaves behind a stub that directs users to `eden du help` if they want more info about the command. This is necessary to ensure the command still appears in the Python help text

# Context

The Python version of `eden du` is extremely behind the Rust version, and it doesn't even seem to work properly on Windows. We should kill it to prevent any confusion.

Reviewed By: lXXXw

Differential Revision: D84772532

fbshipit-source-id: 7167b100143e2f2f53edbe98ba8ffed1fd9f06d8