Add warning icon to master bookmark
Summary:
Even though are now showing master in recommended bookmarks mode, we should include some indication that this not the recommended place to work off/rebase onto. An idea was to add a warning icon next to master with a tooltip.

Open to feedback about this, including if the icon/tooltip should also be in the bookmarks manager list as well.

Reviewed By: evangrayk

Differential Revision: D84871006

fbshipit-source-id: 5c37b06e3da76d4d8c0b5a0935f21a4bd8f599cc