purge: make purge operation --abort-on-err
Summary:
With the `--abort-on-err` option, ISL is able to render operation failures correctly. Otherwise, the purge command always returns 0, even if there are errors, making ISL think it's successful.

One of the cases where `sl purge` fails is purging submodules. It is unsupported for now and users are expected to do it manually, mainly because removing the submodule directories doesn't necessarily undo all the changes from pulling untracked submodules.

Reviewed By: evangrayk

Differential Revision: D84543330

fbshipit-source-id: aa5f2265472609a2484d243d8443cf7ed6abacb1