Change t() to <T> in TopBar.tsx
Summary:
While working on this diff stack I've found `import {t} from './i18n'` did not play well — Prettier would keep deleting the line or changing it to `import {T}` on format even if it was being used within the file.

Since it looks like `<T />` provides the same functionality and is used by other components in the codebase, this diff replaces the import from `t` to `T`.

Reviewed By: evangrayk

Differential Revision: D84414663

fbshipit-source-id: 4564271b9ab7acedaf6c5e66509a77689cf34e65