cgdm: add facet for grouping changesets
Summary: Add a repo facet for grouping changesets into components. For repos that don't have CGDM enabled we use the dummy implementation that returns the ungrouped changesets as they are. For repos that do have CGDM enabled we group the changesets into their components, and for full components we include the IDs pointing to the GDM entries for the component and the commit pack file items.

Reviewed By: RajivTS

Differential Revision: D84921689

fbshipit-source-id: b2f73d79c832bfcafd9ac21f6951afba69638c27