git/protocol: modify the git protocol code to make use of CGDM
Summary: Modify the git protocl code for fetch requests to make use of CGDM. Repos that don't use CGDM will use `DummyCgdmChangesetDivider` which will treat all changesets individually and so we should get the same behaviour. For repos that have CGDM enabled we use the CGDM preloaded blob to group changesets into components, and fetch compacted blobs for every full component to reduce the total number of blobs fetched. This provided a speed up for requests that require fetching data for a large number of changesets.

Reviewed By: RajivTS

Differential Revision: D84921690

fbshipit-source-id: d3084a0becb9dab70dcae968eb62c9cdbfc61642