Support hiding comments
Summary:
The inline comment webviews support hiding individual comments, and we want to support this on the group level as well. I added a button to toggle visibillity for an entire comment group. The state is synced both ways, too -- hiding all of the comments in a particular group will mark the group as hidden as well.

This uses some of the existing logic to apply a "hidden" badge to hidden comments, which I applied to the comment group as well. I opted not to show the group as "disabled" (i.e. with reduced opacity) because this already existed, and  I'm planning to reserve that for resolved comments.

Reviewed By: timrc

Differential Revision: D84729334

fbshipit-source-id: ee57c3be451706df36a4eb6ed95a5e89a7300378