Extend INSTRUMENT_THRIFT_CALL to accept RequestContext
Summary:
Coroutines are unable to work with thread level storage. This is due to coroutines suspending and resuming on any thread on an executor.

For thrift requests, an alternative is to use the RequestParam and RequestContext types. These contain the same data that the futures-based thread level storage contained.

This diff allows the client pid to be captured by this thrift RequestContext instead of only from thread level storage.

Reviewed By: MichaelCuevas

Differential Revision: D84889940

fbshipit-source-id: 32f1b02d2c95542a1cd2bce79fba6e132613ee13