Remove DangerousOverride
Summary:
DangerousOverride is a way to change the RepoObject oncealready created, essentially overriding the config. In this particular case it was used for `dry
_run` of repo import, essentially keeping an in memory view of the repo that did not correspond with the prod view.

That being said, we are trying to remove uses of it (D61391527). Instead of moving to build the initial repo with the desired config, I'm deleting the option to `dry_run`, according to pierrec:
```
I think it would be OK to remove it since we haven't used it for a long time,
and this tool should be something we can get rid of too soon enough.
We could re-implement dry-run differently, by replacing any higher level call
that would write things with a print if we really cared
```

Differential Revision: D85055633

fbshipit-source-id: d9d1987897db0f9723f7fc3cc6808e07e551332a