Support different loggers depending on mode + helper scripts
Summary:
We want to be able to run incremental and full executions of the repo metadata logger, pointing to different Loggers, V2 and V2Full.

This diff abstracts the different loggers in an enum, using a different variant accoring to the execution mode.

Also includes helper scripts (generated by claude) to support manual execution.

Reviewed By: RajivTS

Differential Revision: D84918450

fbshipit-source-id: bc7b0db86bae298b8357d7509f685e060c4cbb17