add blockWithCancel to fault injector
Summary:
The fault that has it all! The test-ification of cancellation! What we've all been waiting for! Drumroll please...
`BlockWithCancel`

Add a new fault injection type that blocks and cooperatively polls a cancellation token. This polling function will also have a timeout that will report a different type of exception, which will help identify in unit tests when we fail to cancel.

Reviewed By: muirdm

Differential Revision: D84203151

fbshipit-source-id: 75eed5b1998a2d9d22453e4e4c00e63b7a8cb4c2