backingstore: fix batch resolution with duplicate keys
Summary:
Fix batch_with_callback to track and resolve things properly when the request contains duplicate keys.

Basically, support tracking a list of request indexes per-key instead of only a single index. Then, if there are multiple indexes for a result item, we fan out the result for each index.

This requires being able to copy the result data as we fan out, so I tweaked the trait to require the data to be Clone.

Reviewed By: MichaelCuevas

Differential Revision: D85160491

fbshipit-source-id: af4285f53c28d1411e071a54b6b6aae9d4675e23