SaplingBackingStore: add prefetchBlobs fast path
Summary:
Add fast path for prefetchBlobs that achieves two major optimizations at once:
1. Sets the sapling IGNORE_RESULT flag which optimizes the blob fetches to check for presence in hgcache, but avoid actually fetching the blob.
2. Directs the batch query directly to the sapling stores instead of going through the SaplingBackingStore queueing. The queueing has a ton of per-item overhead (at least 5 per-item mutex acquisitions and various copies/allocations), and doesn't "make sense" to use when you already have a batch of ids in-hand. The queueing's purpose is to batch up unary fetches automatically (but in this case we already have a batch). Plus, going through the queue actually destroys the nice big 20k batch - when dequeueing, the largest blob batch we do is only 32!

Using the IGNORE_RESULT flag here is safer than the previous attempt because the blob fetches are not going through the queue, so we don't have to worry about collapsing an IGNORE_RESULT fetch for a blob with an unrelated simultaneous fetch that wants the result.

I tried to maintain all the metrics/tracebuses/log messages that made sense.

The new behavior is controlled by experimental.ignore-prefetch-result (default on) and experimental.prefetch-optimizations (default off).

Reviewed By: MichaelCuevas

Differential Revision: D84871666

fbshipit-source-id: 671dad17d7a2a86ba06cb9ba1e3f3249761cc8d4