glob: force parallelism for recursive globs
Summary:
Add aysnc points to recursive glob evaluation so it can saturate the executor, using multiple threads/cores.

Basically, the first N (defaults to 3) depths of `**` glob evaluation are now forced through the executor so the beginning of the recursive glob evaluation gets scheduled on multiple threads in the executor. This behavior can be disabled by turning the threshold down to 0.

The new behavior is also gated on the unified experimental.prefetch-optimizations flag (default off).

When the hgcache is warm, ObjectStore->getTree() always returns ready values, synchronously. In this case, recursive (i.e. `**`) glob evaluation operates single threaded on one core since the only thing it does is call getTree(). This hinders performance since we could be using a lot of cores to traverse in parallel.

Reviewed By: MichaelCuevas

Differential Revision: D84871667

fbshipit-source-id: 88a3fcf9e2341dab0889bf617c6c3bb71f09024a