HgProxyHash: add async point to getBatch
Summary:
If we get a lot of keys, use makeNotReadyImmediateFuture to ensure we go through the executor. This avoids blocking when processing a large list of ids.

In particular, "eden prefetch" currently does this:

```
giant_list_of_files = [...]

futures = []
for chunk in  giant_list_of_files.chunks(20_000):
  futures.push(store->prefetch_blobs(chunk))

collect(futures)
```

If giant_list_of_files has millions of things in it, we spend a significant amount of time doing the blocking HgProxyHash::getBatch call synchronously before we even get to the `collect` phase. With this improvement the HgProxyHash::getBatch work will be asynchronous (when fetching a big amount of files).

This new behavior is gated by the experimental.prefetch-optimizations flag (default off).

Reviewed By: MichaelCuevas

Differential Revision: D84871668

fbshipit-source-id: 1ff6987d6b161409c5bdbca7e318276d63bd8db8