TreeCache: add support for ShardedLruCache
Summary:
Make TreeCache able to use ShardedLruCache instead of ObjectCache. The TreeCache doesn't use the ObjectCache's fancy InterestHandle stuff, so this is relatively easy compared to changing the BlobCache.

I noticed a lot of contention on the TreeCache during a highly parallel "eden prefetch". xavierd added this nifty sharded cache, so let's use it.

The ShardedLruCache mode is activated when the treecache.shards config is greater than 0 (default 32 like metadata cache) and the experimental.prefetch-optimizations config is true (default false).

I added basic stats support for cache size and byte size, along with hits and misses.

Reviewed By: MichaelCuevas

Differential Revision: D84871665

fbshipit-source-id: 1ea242c786e0a7b0b5a46356022e3b4e9b0299d4