glob: don't populate ResultList when suppressFileList=true
Summary:
Avoid locking and updating the result list if we aren't going to use the result list.

When prefetching blobs, we don't need to propagate the glob result. We already had the suppressFileList flag, but the results were suppressed at the end, after doing the work to collect them. Instead, we can skip collecting the results as we go.

I used a pointer for optionality which matches the `PrefetchList* fileBlobsToPrefetch` arg. Maybe they should use `optional<reference_wrapper>` - not sure.

This optimization is gated behind the experimental.prefetch-optimizations flag.

Reviewed By: genevievehelsel

Differential Revision: D84871678

fbshipit-source-id: c9415c6442478ee54a7fc824c5be16c1bbee4e6b