storemodel: switch Box<dyn TreeEntry> to Arc<dyn TreeEntry>
Summary:
I want to make cheap clones of TreeEntry in the backingstore, so let's swap Box to Arc and mark TreeEntry as Sync.

In particular, I want to handle the case when the same Key is requested multiple times in a single batch. We need to be able to copy the data as we fan out resolving the duplicate requests.

I considered some other options such as implementing `Clone` for `Box<dyn TreeEntry>` or passing `&dyn TreeEntry` to the callback, but switching to Arc was simple enough, and I like the property that all the backingstore data types are `Clone` now.

Reviewed By: MichaelCuevas

Differential Revision: D85160489

fbshipit-source-id: 4e08100cfcc1af759ad036acc6fd3ec6e93b88e2