SaplingBackingStore: re-introduce support for IGNORE_RESULT flag
Summary:
Add the ability for eden to opt-in to IGNORE_RESULT when batch fetching blobs. Nothing currently opts in yet.

The IGNORE_RESULT flag tells the sapling scmstore to make sure the blob is present in cache, but skip actually fetching the blob. It speeds things up when prefetching blobs from a warm hgcache.

I originally tried this in D80666100, but reverted in D82125320 due to crashes when cause=Prefetch items collided with cause=Unknown items in the queue. This time I have a much better, safer approach (see D84871666).

Reviewed By: zzl0

Differential Revision: D84871675

fbshipit-source-id: 43dbaea30dfe3f6863fd76e3206aa7f9db899cef