glob: add option to skip dedupe of blobs to prefetch
Summary:
We spend quite a lot of time deduping the list of files when there are millions of files - upwards of 4s. It doesn't seem like there will be that many duplicates - only separate files that happen to have the same content (and in hg repos it will be rarer due to filenode p1+p2). Plus, the SaplingBackingStore will actually dedupe fetches by HgId within the same batch, so I don't think there is much risk even if we have big duplicate files.

I left a config flag to re-enable, just in case. Currently this is also gated behind the experimental.prefetch-optimizations config.

Reviewed By: MichaelCuevas

Differential Revision: D84871674

fbshipit-source-id: 4be2d23273d9e156d288e344a98315f4b892ec3e