Add retries to the diff_service when we get a "repo not found" error
Summary:
Some of the queries to the diff_service get a repo not found error ocasionally:

```
diff_error(DiffError { reason: "repo not found: xrepo_test_large" })
```

This happens when the query goes to a task that doesn't contain the shard for that repo. This could happen due to bad timing while shard manager is reallocating shards (needs more investigation), but in those cases it's safe to retry and find a task with the right details.

NOTE: Ideally we will want a proper error type for this condition instead of matching on strings. This will come on a coming diff.

Reviewed By: chiara125

Differential Revision: D85163727

fbshipit-source-id: 0b2a1a5d942c872f6d63624bc816377474d7f14c