Add config for auto migration to FilteredFS
Summary:
Add edenfs config to control whether the user should be migrated to filteredfs when starting edenfs.

## Context Overview
A two-phase migration has been planned to enforce the FilteredFS feature on selected hosts.
### Phase 1 - EdenFS
Update files which EdenFS relies on when creating a backing store and files sapling checks to decide if FilteredFS should be enabled.
The files include:
- `.hg/sparse`
- `.hg/requires`
- `client_dir/config.toml`
- `client_dir/SNAPSHOT`

After all these files updated, a marker file(`.hg/edensparse_migration`) will be created to mark the completion of this phase.

### Phase 2 - Sapling
These operations will be done when Sapling
- Config sapling to enable `extensions.edensparse` and disable `extensions.sparse`.
- On the next checkout operation if there exists the marker file created by EdenFS in Phase 1, apply the filter to the current checkout.

Reviewed By: MichaelCuevas

Differential Revision: D84863597

fbshipit-source-id: 28de115143a4b60c589068d567e58bd0107227cd