InodeMap: don't reload experimental.lazy-inode-persistence config
Summary: This config knob does not make sense to hot reload while eden is running since it would leave TreeInodes and the overlay in an inconsistent state. In particular, if the config is disabled for a running eden, it could cause eden to never persist inode numbers for directories that are currently expecting to be lazily perissted later.

Reviewed By: kavehahmadi60

Differential Revision: D85265500

fbshipit-source-id: 4750481f5870bc085f7e56f99d6a18ad5de68f42