edenfs-client: add pre/post checkout callbacks to perform filter migration
Summary:
# This diff

Adds pre/post checkout callbacks that are performed during non-dry-run EdenFS checkout operations.

NOTE: It's not possible to apply a filter during a checkout dry run, so we have to wait until the "Normal" checkout operation before attempting the migration. This technically means that a user's checkout operation could race and get a non-conflicting dry run and then a conflicting/failed Normal checkout operation afterwards. This is always a possibility (even without migration), and it's not really possible to avoid if we want to migrate users to FilteredFS transparently.

Adds a few more test cases as well:

- Adds distinct cases for dirty status in filtered directory and dirty status outside filtered directory
- Adds a test for rebase
- adds tests for checking out to a different commit (not just checking out to the current commit)

# Context

See D85097943.

Reviewed By: muirdm

Differential Revision: D85099845

fbshipit-source-id: cf8726c3ee87d24e056568e2271fcf9a0a9966d2