Make blobstore monad subcommands build a RepoBlobstore
Summary:
Until this diff, blobstore monad subcommands were effectively building a RepoBlobstore, composed of exactly the same parts as a RepoBlobstore, but not through RepoBlobstore methods.

Later in the stack, this would cause issues as the blobstore created is not a KeyedBlobstore as it's not a RepoBlobstore, although it's constructed almost the same way.

Let's build it as a RepoBlobstore

Reviewed By: YousefSalama

Differential Revision: D85070417

fbshipit-source-id: 62e423a1e42781a24e5241dffa5821f80748d61e