errors: introduce NETWORK_ERROR into thrift error types
Summary:
Introduce NETWORK_ERROR as a new error type into EdenFS. The errorCode field will be used as network error code.

# Context

As a FUSE system that fetches trees & blobs remotely, it makes sense for Eden to have network error as an error category. It's also helpful for users to know the errors are from the networking layer rather than other layers. In fact network errors account for the majority of Eden errors that buck2 sees as of today.

NOTE: Since network errors may be from various networking libraries (e.g., libcurl, http), there is chance of code collisions from different libraries. This is not a current problem, but we can utilize higher bits of the error code to identify library should this become an issue.

Reviewed By: kavehahmadi60

Differential Revision: D85065711

fbshipit-source-id: 9bd073593e3c4bba4e87ca5997afed90bb139406