Migrate away from ScubaJson in favour of TulipV2
Summary:
In the parent diff, we introduced types to represent the `mononoke_bookmark` and `mononoke_git_content_ref` scribe categories.

Now, replace the parsing logic to use these thrift types to deserialize from TulipV2 format instead of using the `ScubaJson` format.

This must happen as `ScubaJson` will be sunset next week.

For integration tests, still serialize and deserialize the scuba data as json so we can parse it with `cat <file> | jq`.

Reviewed By: YousefSalama

Differential Revision: D85240900

fbshipit-source-id: a32a5279db890b11a30f08992302e4b435546487