filters: sort paths before creating v1 FilterIds
Summary:
# This diff

When multiple filters are enabled at once, the paths are sorted prior to serialization. This ensures we have a stable serialization of filter paths, and it allows FilterId calculation to be stable across separate `sl` operations.

NOTE: this is only safe to do because no users are currently using more than 1 active filter. Therefore, it's not possible for any user to have an unsorted FilterId active and then be stuck with that FilterId when this change lands.

# Context

Prior to sorting filters, it was possible that filter path ordering would change between `sl` invocations, and therefore FilterId calculation would change slightly. When the FilterId hash changed, Sapling would pass the new FilterId to EdenFS, and Eden would get confused (thinking a different FilterId should be active) and fail any `getScmStatusV2` calls with the following error:

```
error.EdenError: error computing status: requested parent commit is out-of-date: requested <hash1> but current parent commit is <hash1>
  Try running `eden doctor` to remediate
```

We can avoid this situation altogehter by ensuring that FilterId calculation is stable across the same active filter paths.

Reviewed By: giorgidze

Differential Revision: D85454747

fbshipit-source-id: 8e541381fc48cc7a785ed5c35c0034e344f4555d